+++ tests/LoggerTest.php
@@ -15,6 +15,7 @@ use Yiisoft\Log\Logger;
 use Yiisoft\Log\Message;
 use Yiisoft\Log\Target;
 use Yiisoft\Log\Tests\TestAsset\DummyTarget;
+use Yiisoft\Log\Tests\TestAsset\StubContextProvider;
 
 use function memory_get_usage;
 
@@ -67,7 +68,7 @@ final class LoggerTest extends TestCase
         $this->assertSame('application', $messages[0]->context('category'));
         $this->assertSame([
             'file' => __FILE__,
-            'line' => 61,
+            'line' => 62,
             'function' => 'log',
             'class' => Logger::class,
             'type' => '->',
@@ -400,6 +401,24 @@ final class LoggerTest extends TestCase
 
         $this->setInaccessibleMessages($logger, [$message]);
         $logger->flush(true);
+    }
+
+    public function testSetTraceLevelWithCustomContextProvider(): void
+    {
+        $logger = new Logger(contextProvider: new StubContextProvider());
+
+        $this->expectException(RuntimeException::class);
+        $this->expectExceptionMessage('"Logger::setTraceLevel()" is unavailable when using a custom context provider.');
+        $logger->setTraceLevel(0);
+    }
+
+    public function testSetExcludedTracePathsWithCustomContextProvider(): void
+    {
+        $logger = new Logger(contextProvider: new StubContextProvider());
+
+        $this->expectException(RuntimeException::class);
+        $this->expectExceptionMessage('"Logger::setExcludedTracePaths()" is unavailable when using a custom context provider.');
+        $logger->setExcludedTracePaths([]);
     }
 
     /**